requirements.txt
asgiref==3.5.1
certifi==2021.10.8
charset-normalizer==2.0.12
distlib==0.3.4
Django==4.0.4
django-cors-headers==3.12.0
django-modeladmin-reorder==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
filelock==3.6.0
idna==3.3
Pillow==9.1.1
platformdirs==2.5.2
psycopg2-binary==2.9.3
PyJWT==2.4.0
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
virtualenv==20.14.1